refactor(ui): fixup paths
